Changes to Css and views Minor fixes to controllers

+++ Controllers/HomeController.cs
@@ -20,19 +20,22 @@ namespace Culinaria.Controllers
     {
         private readonly Uri baseAdress = new Uri("https://localhost:44328/api/recipes");
 
-        private readonly ILogger<HomeController> _logger;
         private readonly UserManager<CulinariaUser> _userManager;
 
-        public HomeController(UserManager<CulinariaUser> userManager, ILogger<HomeController> logger)
+        public HomeController(UserManager<CulinariaUser> userManager)
         {
-            _logger = logger;
             _userManager = userManager;
 
         }
 
         [AllowAnonymous]
-        public async Task<IActionResult> IndexAsync(string? searchBar)
+        public async Task<IActionResult> IndexAsync(string? searchBar, int? n)
         {
+            if (n == 0)
+            {
+                ViewBag.StatusMessage = "Receita criada com sucesso!";
+            }
+
             List<Recipe> list = null;
             using (var client = new HttpClient())
             {
@@ -50,6 +53,10 @@ namespace Culinaria.Controllers
                 {
                     list = await response.Content.ReadAsAsync<List<Recipe>>();
                 }
+                else
+                {
+                    return BadRequest();
+                }
 
             }
             
@@ -57,8 +64,12 @@ namespace Culinaria.Controllers
         }
 
         [AllowAnonymous]
-        public async Task<IActionResult> Recipe(int id)
+        public async Task<IActionResult> Recipe(int id, int? n)
         {
+            if (n == 0)
+            {
+                ViewBag.StatusMessage = "Receita editada com sucesso!";
+            }
             Recipe recipe = null;
             using (var client = new HttpClient())
             {
@@ -69,12 +80,12 @@ namespace Culinaria.Controllers
                 HttpResponseMessage response = await client.GetAsync(client.BaseAddress + "/" + id);
                 if (response.IsSuccessStatusCode)
                 {
-                    //var resp = response.Content.ReadAsStringAsync();
-                    //resp.Wait();
-                    //recipe = JsonConvert.DeserializeObject<Recipe>(resp.Result);
-                    //recipe = await response.Content.ReadAsAsync<Recipe>();
                     var responseDoc = response.Content.ReadAsStringAsync().Result;
                     recipe = JsonConvert.DeserializeObject<Recipe>(responseDoc);
+                }
+                else
+                {
+                    return BadRequest();
                 }
             }
 
@@ -83,15 +94,11 @@ namespace Culinaria.Controllers
                 return NotFound();
             }
 
-
-
             return View(recipe);
         }
 
-        public async Task<IActionResult> CreateAsync()
+        public IActionResult Create()
         {
-            var user = await _userManager.GetUserAsync(User);
-            ViewBag.userId = user.Id;
             return View();
         }
 
@@ -102,22 +109,31 @@ namespace Culinaria.Controllers
 
         public async Task<IActionResult> Delete(int id)
         {
-            using (var client = new HttpClient())
+            if (!await IsOwnerAsync(id))
             {
-                client.BaseAddress = this.baseAdress;
-                HttpResponseMessage response = await client.DeleteAsync(client.BaseAddress + "/" + id);
-                if (response.IsSuccessStatusCode)
-                {
-                    return RedirectToAction("Index");
-                }
-
-                return View();
-                
+                return Forbid();
             }
+            using (var client = new HttpClient())
+                {
+                    client.BaseAddress = this.baseAdress;
+                    HttpResponseMessage response = await client.DeleteAsync(client.BaseAddress + "/" + id);
+                    if (response.IsSuccessStatusCode)
+                    {
+                        return RedirectToAction("Index");
+                    }
+
+                    return View();
+
+                }
         }
 
         public async Task<IActionResult> Edit(int id)
         {
+            if (!await IsOwnerAsync(id))
+            {
+                return Forbid();
+            }
+
             Recipe recipe = null;
             using (var client = new HttpClient())
             {
@@ -139,10 +155,7 @@ namespace Culinaria.Controllers
                 return NotFound();
             }
 
-            if (!(user.Id == recipe.OwnerId) )
-            {
-                return Forbid();
-            }
+            
 
             return View(recipe);
         }
@@ -151,6 +164,28 @@ namespace Culinaria.Controllers
         public IActionResult Error()
         {
             return View(new ErrorViewModel { RequestId = Activity.Current?.Id ?? HttpContext.TraceIdentifier });
+        }
+
+
+        private async Task<bool> IsOwnerAsync(int recipeId)
+        {
+            var user = await _userManager.GetUserAsync(User);
+            string ownerId = null;
+            using (var client = new HttpClient())
+            {
+                client.BaseAddress = this.baseAdress;
+                client.DefaultRequestHeaders.Clear();
+                //Define request data format  
+                client.DefaultRequestHeaders.Accept.Add(new MediaTypeWithQualityHeaderValue("application/json"));
+                HttpResponseMessage response = await client.GetAsync(client.BaseAddress + "/" + recipeId);
+                if (response.IsSuccessStatusCode)
+                {
+                    var responseDoc = response.Content.ReadAsStringAsync().Result;
+                    ownerId = JsonConvert.DeserializeObject<Recipe>(responseDoc).OwnerId;
+                }
+            }
+
+            return user.Id == ownerId;
         }
     }
 }
